refactor: move query view generation to subscription scope
