+++ packages/jazz-tools/src/tools/coValues/coList.ts
@@ -1,12 +1,5 @@
 import { calcPatch } from "fast-myers-diff";
-import type {
-  CoID,
-  CoValueUniqueness,
-  JsonValue,
-  RawCoID,
-  RawCoList,
-  RawCoMap,
-} from "cojson";
+import type { CoValueUniqueness, JsonValue, RawCoID, RawCoList } from "cojson";
 import { cojsonInternals } from "cojson";
 import {
   accessChildByKey,
@@ -31,10 +24,8 @@ import {
   ItemsSym,
   loadCoValueWithoutMe,
   makeRefs,
-  unstable_mergeBranch,
   parseCoValueCreateOptions,
   parseSubscribeRestArgs,
-  queryModifierFields,
   Ref,
   RefEncoded,
   RefsToResolve,
@@ -47,11 +38,7 @@ import {
   subscribeToExistingCoValue,
   SubscribeListenerOptions,
   SubscribeRestArgs,
-  SubscriptionScope,
   TypeSym,
-  type WhereClause,
-  WhereComparisonOperators,
-  WhereLogicalOperators,
 } from "../internal.js";
 
 /**
@@ -521,11 +508,6 @@ export class CoList<out Item = any>
 type CoListItem<L> = L extends CoList<unknown> ? L[number] : never;
 
 export class CoListJazzApi<L extends CoList> extends CoValueJazzApi<L> {
-  private cachedQueryView: {
-    queryView: Record<number, number>;
-    rawLength: number;
-  } | null = null;
-
   constructor(
     private coList: L,
     private getRaw: () => RawCoList,
@@ -861,76 +843,10 @@ export class CoListJazzApi<L extends CoList> extends CoValueJazzApi<L> {
   }
 
   /**
-   * Returns the query modifiers ($where, $orderBy, $limit, $offset) used to load the CoList
-   * @internal
-   */
-  private get queryModifiers() {
-    return this._subscriptionScope?.queryModifiers ?? {};
-  }
-
-  private evaluateWhereClause(
-    whereClause: WhereClause,
-    rawIndex: number,
-    fieldAccessors: Record<string, Record<number, any>>,
-  ): boolean {
-    if ("field" in whereClause) {
-      const valueToFilter = fieldAccessors[whereClause.field]?.[rawIndex];
-      return this.evaluateFieldCondition(
-        whereClause.operator,
-        valueToFilter,
-        whereClause.value,
-      );
-    } else {
-      switch (whereClause.combinator) {
-        case WhereLogicalOperators.$and:
-          return whereClause.conditions.every((condition) =>
-            this.evaluateWhereClause(condition, rawIndex, fieldAccessors),
-          );
-        case WhereLogicalOperators.$or:
-          return whereClause.conditions.some((condition) =>
-            this.evaluateWhereClause(condition, rawIndex, fieldAccessors),
-          );
-        case WhereLogicalOperators.$not:
-          return !this.evaluateWhereClause(
-            whereClause.conditions[0]!,
-            rawIndex,
-            fieldAccessors,
-          );
-      }
-    }
-  }
-
-  private evaluateFieldCondition(
-    operator: string,
-    valueToFilter: any,
-    value: any,
-  ): boolean {
-    switch (operator) {
-      case WhereComparisonOperators.$eq:
-        return valueToFilter === value;
-      case WhereComparisonOperators.$ne:
-        return valueToFilter !== value;
-    }
-    if (valueToFilter === undefined) return false;
-    switch (operator) {
-      case WhereComparisonOperators.$gt:
-        return valueToFilter > value;
-      case WhereComparisonOperators.$gte:
-        return valueToFilter >= value;
-      case WhereComparisonOperators.$lt:
-        return valueToFilter < value;
-      case WhereComparisonOperators.$lte:
-        return valueToFilter <= value;
-      default:
-        return true;
-    }
-  }
-
-  /**
    * Get the raw indexes of the CoList items that are loaded and accessible.
    * @internal
    */
-  private accessibleRawIndexes(): number[] {
+  accessibleRawIndexes(): number[] {
     const rawIndexes: number[] = [];
     const rawEntries = this.raw.entries();
     for (let rawIndex = 0; rawIndex < rawEntries.length; rawIndex++) {
@@ -987,55 +903,7 @@ export class CoListJazzApi<L extends CoList> extends CoValueJazzApi<L> {
    * @internal
    */
   private get queryView(): Record<number, number> | null {
-    if (Object.keys(this.queryModifiers).length === 0) {
-      return null;
-    }
-    const rawLength = this.raw.entries().length;
-    // TODO we should invalidate the cached query view if the CoList elements change
-    if (this.cachedQueryView && this.cachedQueryView.rawLength === rawLength) {
-      return this.cachedQueryView.queryView;
-    }
-    const {
-      limit = Infinity,
-      offset = 0,
-      orderBy,
-      where,
-    } = this.queryModifiers;
-    const queriedFields = queryModifierFields(this.queryModifiers);
-    const fieldAccessors = Object.fromEntries(
-      queriedFields.map((field) => [field, this.fieldAccessor(field)]),
-    );
-    let filteredArrayIndexes = this.accessibleRawIndexes();
-    if (where) {
-      filteredArrayIndexes = filteredArrayIndexes.filter((rawIndex) =>
-        this.evaluateWhereClause(where, rawIndex, fieldAccessors),
-      );
-    }
-    let sortedArrayIndexes = filteredArrayIndexes;
-    if (orderBy?.length) {
-      sortedArrayIndexes = sortedArrayIndexes.toSorted((rawIdxA, rawIdxB) => {
-        for (const { field, orderDirection } of orderBy) {
-          const dir = orderDirection === "desc" ? -1 : 1;
-          const aValue = fieldAccessors[field]?.[rawIdxA];
-          const bValue = fieldAccessors[field]?.[rawIdxB];
-
-          // Undefined values go last, regardless of order direction
-          if (aValue === undefined && bValue === undefined) continue;
-          if (aValue === undefined) return 1;
-          if (bValue === undefined) return -1;
-
-          if (aValue < bValue) return -1 * dir;
-          if (aValue > bValue) return 1 * dir;
-        }
-        return 0;
-      });
-    }
-    const paginatedArrayIndexes = sortedArrayIndexes
-      .slice(offset, offset + limit)
-      .map((rawIndex, idx) => [idx, rawIndex]);
-    const view = Object.fromEntries(paginatedArrayIndexes);
-    this.cachedQueryView = { queryView: view, rawLength };
-    return view;
+    return this._subscriptionScope?.queryView ?? null;
   }
 
   /**
